pyproject.toml
[tool.poetry]
name = "auto-parliament"
version = "0.1.0"
description = "LLM moral parliament experiments"
authors = ["aaron-sandoval <32021231+aaron-sandoval@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
pyautogen = "^0.3.0"
pandas = "^2.2.2"
matplotlib = "^3.9.2"
datasets = ">=2.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
